fix: update dependencies
