+++ aform/package.json
@@ -44,31 +44,31 @@
 	"dependencies": {
 		"@stonecrop/themes": "workspace:*",
 		"@stonecrop/utilities": "workspace:*",
-		"@vueuse/core": "^12.0.0",
-		"@vueuse/components": "^12.0.0",
-		"vue": "^3.5.11"
+		"@vueuse/core": "^12.7.0",
+		"@vueuse/components": "^12.7.0",
+		"vue": "^3.5.13"
 	},
 	"devDependencies": {
-		"@microsoft/api-documenter": "^7.26.2",
-		"@rushstack/heft": "^0.68.6",
+		"@microsoft/api-documenter": "^7.26.10",
+		"@rushstack/heft": "^0.68.18",
 		"@stonecrop/atable": "workspace:*",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
-		"@vitest/coverage-istanbul": "^2.1.1",
-		"@vitest/ui": "^2.1.1",
+		"@vitejs/plugin-vue": "^5.2.1",
+		"@vitest/coverage-istanbul": "^3.0.6",
+		"@vitest/ui": "^3.0.6",
 		"@vue/test-utils": "^2.4.6",
-		"cypress": "^13.11.0",
+		"cypress": "^14.0.3",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-vue": "^9.11.1",
 		"eslint": "^8.40.0",
-		"jsdom": "^25.0.0",
+		"jsdom": "^26.0.0",
 		"stonecrop-rig": "workspace:*",
 		"typescript": "~5.6.3",
-		"vite-plugin-lib-inject-css": "^2.1.1",
-		"vite": "^5.4.5",
-		"vitest": "^2.1.1",
-		"vue-router": "^4.4.0"
+		"vite-plugin-lib-inject-css": "^2.2.1",
+		"vite": "^6.1.1",
+		"vitest": "^3.0.6",
+		"vue-router": "^4.5.0"
 	},
 	"peerDependencies": {
 		"@stonecrop/atable": "workspace:*"
