+++ beam/package.json
@@ -40,31 +40,31 @@
 		"test:ui": "vitest --ui"
 	},
 	"dependencies": {
-		"@vueuse/core": "^12.0.0",
-		"mqtt": "^5.10.1",
+		"@vueuse/core": "^12.7.0",
+		"mqtt": "^5.10.3",
 		"onscan.js": "^1.5.2",
-		"vue": "^3.5.11"
+		"vue": "^3.5.13"
 	},
 	"devDependencies": {
-		"@microsoft/api-documenter": "^7.26.2",
-		"@rushstack/heft": "^0.68.6",
+		"@microsoft/api-documenter": "^7.26.10",
+		"@rushstack/heft": "^0.68.18",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
-		"@vitest/coverage-istanbul": "^2.1.1",
-		"@vitest/ui": "^2.1.1",
+		"@vitejs/plugin-vue": "^5.2.1",
+		"@vitest/coverage-istanbul": "^3.0.6",
+		"@vitest/ui": "^3.0.6",
 		"@vue/test-utils": "^2.4.6",
-		"cypress": "^13.11.0",
+		"cypress": "^14.0.3",
 		"eslint": "^8.40.0",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-vue": "^9.11.1",
 		"sass": "^1.75.0",
 		"stonecrop-rig": "workspace:*",
 		"typescript": "~5.6.3",
-		"vite-plugin-lib-inject-css": "^2.1.1",
-		"vite": "^5.4.5",
-		"vitest": "^2.1.1",
-		"vue-router": "^4.4.0"
+		"vite-plugin-lib-inject-css": "^2.2.1",
+		"vite": "^6.1.1",
+		"vitest": "^3.0.6",
+		"vue-router": "^4.5.0"
 	},
 	"publishConfig": {
 		"access": "public"
