+++ desktop/package.json
@@ -39,23 +39,23 @@
 		"@stonecrop/atable": "workspace:*",
 		"@stonecrop/stonecrop": "workspace:*",
 		"@stonecrop/themes": "workspace:*",
-		"vue": "^3.5.11"
+		"vue": "^3.5.13"
 	},
 	"devDependencies": {
-		"@microsoft/api-documenter": "^7.26.2",
+		"@microsoft/api-documenter": "^7.26.10",
 		"@miragejs/graphql": "^0.1.13",
-		"@rushstack/heft": "^0.68.6",
+		"@rushstack/heft": "^0.68.18",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
+		"@vitejs/plugin-vue": "^5.2.1",
 		"eslint": "^8.40.0",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-vue": "^9.11.1",
-		"miragejs": "^0.1.47",
+		"miragejs": "^0.1.48",
 		"stonecrop-rig": "workspace:*",
 		"typescript": "~5.6.3",
-		"vite": "^5.4.5",
-		"vue-router": "^4.4.0"
+		"vite": "^6.1.1",
+		"vue-router": "^4.5.0"
 	},
 	"publishConfig": {
 		"access": "public"
