+++ examples/beam/vite.config.ts
@@ -5,6 +5,12 @@ import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vite'
 
 export default defineConfig({
+	server: {
+		fs: {
+			// Allow reading files from the Stonecrop root workspace
+			allow: ['../..'],
+		},
+	},
 	plugins: [vue()],
 	histoire: {
 		plugins: [HstVue()],
