+++ examples/package.json
@@ -32,29 +32,29 @@
 		"@stonecrop/stonecrop": "workspace:*",
 		"@stonecrop/themes": "workspace:*",
 		"@miragejs/graphql": "^0.1.13",
-		"miragejs": "^0.1.47",
-		"pinia": "^2.3.0",
-		"vite": "^5.4.5",
-		"vue": "^3.5.11",
+		"miragejs": "^0.1.48",
+		"pinia": "^3.0.1",
+		"vite": "^6.1.1",
+		"vue": "^3.5.13",
 		"vue-toast-notification": "^3",
 		"xstate": "^4.38.3"
 	},
 	"devDependencies": {
-		"@histoire/plugin-vue": "^0.17.17",
-		"@monaco-editor/loader": "^1.3.2",
+		"@histoire/plugin-vue": "^1.0.0-alpha.2",
+		"@monaco-editor/loader": "^1.5.0",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
-		"@vue-flow/core": "^1.37.1",
-		"cypress": "^13.11.0",
+		"@vitejs/plugin-vue": "^5.2.1",
+		"@vue-flow/core": "^1.42.1",
+		"cypress": "^14.0.3",
 		"eslint": "^8.40.0",
 		"eslint-plugin-vue": "^9.11.1",
-		"histoire": "^0.17.17",
-		"immutable": "^4.3.0",
-		"monaco-editor": "^0.34.1",
+		"histoire": "^1.0.0-alpha.2",
+		"immutable": "^5.0.3",
+		"monaco-editor": "^0.52.2",
 		"portal-vue": "^3.0.0",
 		"typescript": "~5.6.3",
-		"vue-router": "^4.4.0"
+		"vue-router": "^4.5.0"
 	},
 	"engines": {
 		"node": ">=20.11.0"
