+++ node_editor/package.json
@@ -40,23 +40,23 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"@vue-flow/core": "^1.37.1",
-		"vue": "^3.5.11",
-		"vue-router": "^4.4.0",
+		"@vue-flow/core": "^1.42.1",
+		"vue": "^3.5.13",
+		"vue-router": "^4.5.0",
 		"xstate": "^4.38.3"
 	},
 	"devDependencies": {
-		"@microsoft/api-documenter": "^7.26.2",
-		"@rushstack/heft": "^0.68.6",
+		"@microsoft/api-documenter": "^7.26.10",
+		"@rushstack/heft": "^0.68.18",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
+		"@vitejs/plugin-vue": "^5.2.1",
 		"eslint": "^8.40.0",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-vue": "^9.11.1",
 		"stonecrop-rig": "workspace:*",
 		"typescript": "~5.6.3",
-		"vite": "^5.4.5"
+		"vite": "^6.1.1"
 	},
 	"publishConfig": {
 		"access": "public"
