+++ rigs/stonecrop-rig/package.json
@@ -9,10 +9,10 @@
 		"_phase:build": ""
 	},
 	"dependencies": {
-		"@microsoft/api-extractor": "^7.48.0",
-		"@rushstack/heft": "^0.68.6",
-		"@rushstack/heft-api-extractor-plugin": "^0.3.60",
-		"@rushstack/heft-typescript-plugin": "^0.6.3",
+		"@microsoft/api-extractor": "^7.50.1",
+		"@rushstack/heft": "^0.68.18",
+		"@rushstack/heft-api-extractor-plugin": "^0.3.68",
+		"@rushstack/heft-typescript-plugin": "^0.6.11",
 		"typescript": "~5.6.3"
 	},
 	"engines": {
