+++ rush.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://developer.microsoft.com/json-schemas/rush/v5/rush.schema.json",
-	"rushVersion": "5.147.0",
-	"pnpmVersion": "9.15.0",
+	"rushVersion": "5.149.1",
+	"pnpmVersion": "9.15.5",
 	"telemetryEnabled": false,
 
 	"approvedPackagesPolicy": {
