+++ themes/package.json
@@ -30,7 +30,7 @@
 		"preview": "vite preview"
 	},
 	"devDependencies": {
-		"vite": "^5.4.5"
+		"vite": "^6.1.1"
 	},
 	"publishConfig": {
 		"access": "public"
