+++ utilities/package.json
@@ -38,23 +38,23 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"@vueuse/core": "^12.0.0",
-		"vue": "^3.5.11"
+		"@vueuse/core": "^12.7.0",
+		"vue": "^3.5.13"
 	},
 	"devDependencies": {
-		"@microsoft/api-documenter": "^7.26.2",
-		"@rushstack/heft": "^0.68.6",
+		"@microsoft/api-documenter": "^7.26.10",
+		"@rushstack/heft": "^0.68.18",
 		"@typescript-eslint/eslint-plugin": "^7.14.1",
 		"@typescript-eslint/parser": "^7.14.1",
-		"@vitejs/plugin-vue": "^5.1.3",
-		"cypress": "^13.11.0",
+		"@vitejs/plugin-vue": "^5.2.1",
+		"cypress": "^14.0.3",
 		"eslint": "^8.40.0",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-vue": "^9.11.1",
 		"stonecrop-rig": "workspace:*",
 		"typescript": "~5.6.3",
-		"vite": "^5.4.5",
-		"vue-router": "^4.4.0"
+		"vite": "^6.1.1",
+		"vue-router": "^4.5.0"
 	},
 	"publishConfig": {
 		"access": "public"
